requirements.txt
Flask
Flask-Login
Flask-Mail
Flask-Session
Pillow
